+++ lib/api_sentinel.dart
@@ -4,3 +4,7 @@ export 'controllers/api_service.dart';
 export 'controllers/error_handler/error_handler.dart';
 export 'models/monitoring/network_monitoring_params.dart';
 export 'models/monitoring/network_monitoring_function.dart';
+export 'widgets/secret_knock_detector.dart';
+export 'widgets/totp_secret_section.dart';
+export 'controllers/access_controller.dart';
+export 'res/all_controller_keys.dart';