+++ lib/res/all_controller_keys.dart
@@ -1,0 +1,3 @@
+class AllControllerKeys {
+  static String accessControllerKey = 'accessControllerKey';
+}